+++ cmd/moombox/main.go
@@ -42,7 +42,7 @@ import (
 )
 
 var (
-	version = "2.4.17"
+	version = "2.4.18"
 	commit  = ""
 )